requirements.txt
asgiref==3.7.2
Django==5.0.3
django-filter==24.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
setuptools==69.2.0
sqlparse==0.4.4
uritemplate==4.1.1